Check if powershell exists on windows

+++ src/AbpDevTools/Commands/EnableNotificationsCommand.cs
@@ -21,6 +21,11 @@ public class EnableNotificationsCommand : ICommand
     {
         if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
         {
+            if (!PowershellExistsInWindows())
+            {
+                throw new CommandException($"Powershell is not installed in your system. Please install it and try again.");
+            }
+
             var process = Process.Start("powershell", "-Command Install-Module -Name BurntToast");
 
             console.RegisterCancellationHandler().Register(() => process.Kill(entireProcessTree: true));
@@ -36,7 +41,7 @@ public class EnableNotificationsCommand : ICommand
             return;
         }
 
-        if(RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
         {
             var options = NotificationConfiguration.GetOptions();
             options.Enabled = true;
@@ -48,5 +53,11 @@ public class EnableNotificationsCommand : ICommand
         }
 
         throw new CommandException($"This operation isn't supported on {RuntimeInformation.OSDescription} currently. :(");
+    }
+
+    public bool PowershellExistsInWindows()
+    {
+        string regval = Microsoft.Win32.Registry.GetValue(@"HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\PowerShell\1", "Install", null).ToString();
+        return regval.Equals("1");
     }
 }
